pyproject.toml
[tool.poetry]
name = "top-github-scraper"
version = "0.1.1.6"
description = "Scrape top GitHub repositories and users based on keyword"
authors = ["khuyentran1401 <khuyentran1476@gmail.com>"]
homepage = "https://github.com/khuyentran1401/top-github-scraper"
repository = "https://github.com/khuyentran1401/top-github-scraper"
keywords = ["GitHub", "github-api", "scraping", "python"]
readme = "README.md"
license = "Apache-2.0"

[tool.poetry.dependencies]
python = ">=3.7.1,<4.0"
requests = "^2.25.1"
bs4 = "^0.0.1"
pandas = "^1.2.2"
rich = "^9.12.0"
python-dotenv = "^0.15.0"
ipython = "^7.21.0"
tqdm = "^4.58.0"

[tool.poetry.dev-dependencies]
pre-commit = "^2.10.1"
black = "^20.8b1"
flake8 = "^3.8.4"
isort = "^5.7.0"
icecream = "^2.1.0"
loguru = "^0.5.3"
pytest = "^6.2.2"

[tool.black]
line-length = 79
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
